reafactor: renamed the domain filenames to start with lower case as they do not export a class.
