+++ packages/client/src/components/canary-executor/CanaryExecutorButtonSection.tsx
@@ -9,7 +9,7 @@ import log from '../../util/LoggerFactory';
 import { kayentaApiService } from '../../services';
 import { boundMethod } from 'autobind-decorator';
 import './CanaryExecutorButtonSection.scss';
-import { CanaryAdhocExecutionRequest } from '../../domain/Kayenta';
+import { CanaryAdhocExecutionRequest } from '../../domain/kayenta';
 
 interface Stores {
   canaryExecutorStore: CanaryExecutorStore;
