+++ packages/client/src/services/LoadCanaryConfigService.ts
@@ -1,6 +1,6 @@
 import { stores } from '../stores';
 import axios from 'axios';
-import { CanaryConfig } from '../domain/Kayenta';
+import { CanaryConfig } from '../domain/kayenta';
 import log from '../util/LoggerFactory';
 import { connect, ConnectedComponent } from '../components/connectedComponent';
 import { RouterProps } from 'react-router';
